Next destination logic seperated but getting looped and trying to implement by viewmodel now

+++ app/src/main/java/com/samrish/driver/models/Trip.kt
@@ -51,9 +51,9 @@ class VehicleList(
 class TripActions(
     var actions: List<String>,
     var nextLocationName : String,
-    var estimatedTime: Double,
+    var estimatedTime: Int,
     var estimatedDistance: Double,
     var travelledDistance: Double,
-    var travelTime: Double
+    var travelTime: Int
 
 )
